# `MemPostings.PostingsForLabelMatching()`: don't hold the mutex while matching