This changes the `MemPostings.PostingsForLabelMatching` implementation to stop holding the read mutex while matching the label values.

We've seen that this method can be slow when the matcher is expensive, that's why we even added a context expiration check.

However, there are critical process that might be waiting on this mutex: writes (adding new series) and compaction (deleting the garbage-collected ones), so we should avoid holding it for a long period of time.

Given that we've copied the values to a slice anyway, there's no need to hold the lock while matching.

Related to https://github.com/prometheus/prometheus/pull/13286