@colega 2024-06-10
There's obviously some performance penalty when benchmarking this method alone:
```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/tsdb/index
                                                                   │    main     │               nohold                │
                                                                   │   sec/op    │   sec/op     vs base                │
MemPostings_PostingsForLabelMatching/labels=1000/matcher=fast-12     37.45µ ± 1%   38.22µ ± 1%   +2.06% (p=0.000 n=10)
MemPostings_PostingsForLabelMatching/labels=1000/matcher=slow-12     12.41µ ± 0%   12.58µ ± 0%   +1.42% (p=0.000 n=10)
MemPostings_PostingsForLabelMatching/labels=10000/matcher=fast-12    392.6µ ± 1%   404.8µ ± 0%   +3.10% (p=0.000 n=10)
MemPostings_PostingsForLabelMatching/labels=10000/matcher=slow-12    253.2m ± 0%   251.0m ± 1%   -0.86% (p=0.004 n=10)
MemPostings_PostingsForLabelMatching/labels=100000/matcher=fast-12   3.897m ± 0%   4.484m ± 8%  +15.05% (p=0.000 n=10)
MemPostings_PostingsForLabelMatching/labels=100000/matcher=slow-12    3.333 ± 0%    3.296 ± 0%   -1.11% (p=0.000 n=10)
geomean                                                              2.904m        2.995m        +3.14%
```

However, what we're optimizing here is the speed of appending a new series to the tsdb (which needs to hold the lock) and the garbage collection: if the benchmark added in https://github.com/prometheus/prometheus/pull/13286 was using this method for concurrent reads, we'd see an improvement there.